fix #2: accessToken 발급 서버에서 처리하는 방식으로 롤백
